requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
asyncpg==0.30.0
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
cachetools==5.5.0
cattrs==24.1.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
courlan==1.3.1
cryptography==43.0.3
dataclasses-json==0.6.7
dateparser==1.2.0
decorator==4.4.2
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.3
frozenlist==1.4.1
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
googleapis-common-protos==1.65.0
googlemaps==4.10.0
gotrue==2.9.3
greenlet==3.1.1
grpcio==1.67.0
grpcio-tools==1.67.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
htmldate==1.9.1
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
jiter==0.6.1
jsonpatch==1.33
jsonpointer==3.0.0
jusText==3.0.1
langchain==0.3.4
langchain-community==0.3.3
langchain-core==0.3.12
langchain-google-community==2.0.1
langchain-openai==0.2.3
langchain-text-splitters==0.3.0
langsmith==0.1.136
lxml==5.3.0
lxml_html_clean==0.3.1
markdown-it-py==3.0.0
marshmallow==3.23.0
mdurl==0.1.2
moviepy==1.0.3
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.52.1
orjson==3.10.10
packaging==24.1
pillow==11.0.0
platformdirs==4.3.6
portalocker==2.10.1
postgrest==0.17.2
proglog==0.1.10
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.2
pywin32==308
PyYAML==6.0.2
qdrant-client==1.12.0
realtime==2.0.6
regex==2024.9.11
requests==2.32.3
requests-cache==1.2.1
requests-toolbelt==1.0.0
rich==13.9.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.41.0
storage3==0.8.2
StrEnum==0.4.15
supabase==2.9.1
supafunc==0.6.2
tenacity==9.0.0
tiktoken==0.8.0
tld==0.13
tqdm==4.66.5
trafilatura==1.12.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==2.2.3
uvicorn==0.32.0
websockets==13.1
yarl==1.16.0
